// types/abi/TodoApp.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */

import type BN from "bn.js";
import type { ContractOptions } from "web3-eth-contract";
import type { EventLog } from "web3-core";
import type { EventEmitter } from "events";
import type {
  Callback,
  PayableTransactionObject,
  NonPayableTransactionObject,
  BlockType,
  ContractEventLog,
  BaseContract,
} from "./types";

export interface EventOptions {
  filter?: object;
  fromBlock?: BlockType;
  topics?: string[];
}

export interface TodoApp extends BaseContract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  ): TodoApp;
  clone(): TodoApp;
  methods: {
    ownerTodoCount(arg0: string): NonPayableTransactionObject<string>;

    todoList(arg0: number | string | BN): NonPayableTransactionObject<{
      taskid: string;
      task: string;
      flag: boolean;
      0: string;
      1: string;
      2: boolean;
    }>;

    todoToOwner(
      arg0: number | string | BN
    ): NonPayableTransactionObject<string>;

    createTodo(_task: string): NonPayableTransactionObject<void>;

    TodoRemove(id: number | string | BN): NonPayableTransactionObject<void>;

    getTodoListByOwner(owner: string): NonPayableTransactionObject<string[]>;
  };
  events: {
    allEvents(options?: EventOptions, cb?: Callback<EventLog>): EventEmitter;
  };
}